LPS-85787 Keep Preston idea of only one tracker :), sadly we can't keep the provider registrar, since APIO, for now, supports only one provider per type. In the future, we will keep the whole Preston implementation.
